Editinig issues
