+++ TagCloud/CloudLayouter/FigurePaths/IFigurePath.cs
@@ -1,6 +1,6 @@
 ﻿using System.Drawing;
 
-namespace TagCloud.FigurePaths
+namespace TagCloud.CloudLayouter.FigurePaths
 {
     public interface IFigurePath
     {
