--- TagCloud/CloudVisualizer/CloudViewConfiguration/IColorWordPicker.cs
@@ -1,10 +1,0 @@
-﻿using System.Drawing;
-using TagCloud.WordsPreprocessing;
-
-namespace TagCloud.CloudVisualizer.CloudViewConfiguration
-{
-    public interface IColorWordPicker
-    {
-        Color GetColor(Word word);
-    }
-}
