+++ TagCloud/CloudVisualizerSpace/CloudViewConfigurationSpace/IColorWordPicker.cs
@@ -1,0 +1,10 @@
+﻿using System.Drawing;
+using TagCloud.WordsPreprocessing;
+
+namespace TagCloud.CloudVisualizerSpace.CloudViewConfigurationSpace
+{
+    public interface IColorWordPicker
+    {
+        Color GetColor(Word word);
+    }
+}
