+++ TagCloud/NameAttribute.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace TagCloud
+{
+    public class NameAttribute : Attribute
+    {
+        public NameAttribute(string name)
+        {
+            Name = name;
+        }
+
+        public string Name { get; }
+    }
+}
