+++ TagCloud/TagCloudTests/TagCloudTests.csproj
@@ -45,6 +45,9 @@
     <Reference Include="GroboContainer, Version=1.2.45.0, Culture=neutral, processorArchitecture=MSIL">
       <HintPath>..\packages\GroboContainer.1.2.45\lib\net45\GroboContainer.dll</HintPath>
     </Reference>
+    <Reference Include="Newtonsoft.Json, Version=9.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
+      <HintPath>..\packages\Newtonsoft.Json.9.0.1\lib\net45\Newtonsoft.Json.dll</HintPath>
+    </Reference>
     <Reference Include="nunit.framework, Version=3.12.0.0, Culture=neutral, PublicKeyToken=2638cd05610744eb, processorArchitecture=MSIL">
       <HintPath>..\packages\NUnit.3.12.0\lib\net45\nunit.framework.dll</HintPath>
     </Reference>
@@ -58,13 +61,17 @@
     <Reference Include="System.Data" />
     <Reference Include="System.Net.Http" />
     <Reference Include="System.Xml" />
+    <Reference Include="YandexMystem.Wrapper, Version=1.0.0.18234, Culture=neutral, processorArchitecture=MSIL">
+      <HintPath>..\packages\YandexMystem.Wrapper.1.0.0.18234\lib\net452\YandexMystem.Wrapper.dll</HintPath>
+    </Reference>
   </ItemGroup>
   <ItemGroup>
     <Compile Include="CloudLayouter\CircularCloudLayouter_Should.cs" />
-    <Compile Include="CloudLayouter\FigurePaths\SpiralShould.cs" />
+    <Compile Include="CloudLayouter\FigurePaths\Spiral_Should.cs" />
+    <Compile Include="CloudVisualizerSpace\CloudVisualizer_Should.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="WordsPreprocessing\TextAnalyzers\RussianLanguageAnalyzerShould.cs" />
+    <Compile Include="WordsPreprocessing\TextAnalyzers\RussianLanguageAnalyzer_Should.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
@@ -75,6 +82,9 @@
       <Project>{1bdfd59f-bb73-413a-9b98-9e852d784a95}</Project>
       <Name>TagCloud</Name>
     </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="mystem.exe" />
   </ItemGroup>
   <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
