+++ TagCloud/TagCloudTests/packages.config
@@ -3,5 +3,7 @@
   <package id="FluentAssertions" version="5.9.0" targetFramework="net472" />
   <package id="GrEmit" version="3.1.1" targetFramework="net472" />
   <package id="GroboContainer" version="1.2.45" targetFramework="net472" />
+  <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net472" />
   <package id="NUnit" version="3.12.0" targetFramework="net472" />
+  <package id="YandexMystem.Wrapper" version="1.0.0.18234" targetFramework="net472" />
 </packages>
