+++ TagCloud/WordsPreprocessing/TextAnalyzers/ITextAnalyzer.cs
@@ -1,11 +1,9 @@
-﻿using System.Collections;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 
 namespace TagCloud.WordsPreprocessing.TextAnalyzers
 {
     public interface ITextAnalyzer
     {
         Word[] GetWords(IEnumerable<string> words, int count);
-        string AnalyzerName { get; }
     }
 }
